requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.30.0
python-multipart>=0.0.5
sentence-transformers>=2.2.0
tf-keras>=2.15.0
pandas>=2.1.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
selenium>=4.0.0
webdriver-manager>=4.0.0
Pillow>=10.0.0
tabula-py>=2.9.0
urllib3>=2.0.0
numpy<2.0.0
torch>=2.0.0
openpyxl>=3.1.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0
starlette>=0.27.0